// src/graphql/schema.types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  DateTime: { input: string; output: string };
};
// ...
export type CheckListItem = {
  checked: Scalars["Boolean"]["output"];
  title: Scalars["String"]["output"];
};
// ...
export enum Role {
  Admin = "ADMIN",
  SalesIntern = "SALES_INTERN",
  SalesManager = "SALES_MANAGER",
  SalesPerson = "SALES_PERSON",
}
// ...
export type Task = {
  checklist: Array<CheckListItem>;
  comments: TaskCommentsConnection;
  commentsAggregate: Array<TaskCommentsAggregateResponse>;
  completed: Scalars["Boolean"]["output"];
  createdAt: Scalars["DateTime"]["output"];
  createdBy: User;
  description?: Maybe<Scalars["String"]["output"]>;
  dueDate?: Maybe<Scalars["DateTime"]["output"]>;
  id: Scalars["ID"]["output"];
  stage?: Maybe<TaskStage>;
  stageId?: Maybe<Scalars["ID"]["output"]>;
  title: Scalars["String"]["output"];
  updatedAt: Scalars["DateTime"]["output"];
  updatedBy?: Maybe<User>;
  users: Array<User>;
  usersAggregate: Array<TaskUsersAggregateResponse>;
};
// ...
export type TaskCommentsAggregateResponse = {
  avg?: Maybe<TaskCommentsAvgAggregate>;
  count?: Maybe<TaskCommentsCountAggregate>;
  groupBy?: Maybe<TaskCommentsAggregateGroupBy>;
  max?: Maybe<TaskCommentsMaxAggregate>;
  min?: Maybe<TaskCommentsMinAggregate>;
  sum?: Maybe<TaskCommentsSumAggregate>;
};

export type TaskCommentsAvgAggregate = {
  id?: Maybe<Scalars["Float"]["output"]>;
};

export type TaskCommentsConnection = {
  /** Array of nodes. */
  nodes: Array<TaskComment>;
  /** Paging information */
  pageInfo: OffsetPageInfo;
  /** Fetch total count of records */
  totalCount: Scalars["Int"]["output"];
};
// ...
export type TaskStage = {
  createdAt: Scalars["DateTime"]["output"];
  createdBy: User;
  id: Scalars["ID"]["output"];
  tasks: Array<Task>;
  tasksAggregate: Array<TaskStageTasksAggregateResponse>;
  title: Scalars["String"]["output"];
  updatedAt: Scalars["DateTime"]["output"];
  updatedBy?: Maybe<User>;
};
// ...
export type TaskStageTasksAggregateGroupBy = {
  completed?: Maybe<Scalars["Boolean"]["output"]>;
  createdAt?: Maybe<Scalars["DateTime"]["output"]>;
  description?: Maybe<Scalars["String"]["output"]>;
  dueDate?: Maybe<Scalars["DateTime"]["output"]>;
  id?: Maybe<Scalars["ID"]["output"]>;
  stageId?: Maybe<Scalars["ID"]["output"]>;
  title?: Maybe<Scalars["String"]["output"]>;
  updatedAt?: Maybe<Scalars["DateTime"]["output"]>;
};

export type TaskStageTasksAggregateResponse = {
  avg?: Maybe<TaskStageTasksAvgAggregate>;
  count?: Maybe<TaskStageTasksCountAggregate>;
  groupBy?: Maybe<TaskStageTasksAggregateGroupBy>;
  max?: Maybe<TaskStageTasksMaxAggregate>;
  min?: Maybe<TaskStageTasksMinAggregate>;
  sum?: Maybe<TaskStageTasksSumAggregate>;
};

export type TaskStageTasksAvgAggregate = {
  id?: Maybe<Scalars["Float"]["output"]>;
  stageId?: Maybe<Scalars["Float"]["output"]>;
};

export type TaskStageTasksCountAggregate = {
  completed?: Maybe<Scalars["Int"]["output"]>;
  createdAt?: Maybe<Scalars["Int"]["output"]>;
  description?: Maybe<Scalars["Int"]["output"]>;
  dueDate?: Maybe<Scalars["Int"]["output"]>;
  id?: Maybe<Scalars["Int"]["output"]>;
  stageId?: Maybe<Scalars["Int"]["output"]>;
  title?: Maybe<Scalars["Int"]["output"]>;
  updatedAt?: Maybe<Scalars["Int"]["output"]>;
};

export type TaskStageTasksMaxAggregate = {
  createdAt?: Maybe<Scalars["DateTime"]["output"]>;
  description?: Maybe<Scalars["String"]["output"]>;
  dueDate?: Maybe<Scalars["DateTime"]["output"]>;
  id?: Maybe<Scalars["ID"]["output"]>;
  stageId?: Maybe<Scalars["ID"]["output"]>;
  title?: Maybe<Scalars["String"]["output"]>;
  updatedAt?: Maybe<Scalars["DateTime"]["output"]>;
};

export type TaskStageTasksMinAggregate = {
  createdAt?: Maybe<Scalars["DateTime"]["output"]>;
  description?: Maybe<Scalars["String"]["output"]>;
  dueDate?: Maybe<Scalars["DateTime"]["output"]>;
  id?: Maybe<Scalars["ID"]["output"]>;
  stageId?: Maybe<Scalars["ID"]["output"]>;
  title?: Maybe<Scalars["String"]["output"]>;
  updatedAt?: Maybe<Scalars["DateTime"]["output"]>;
};

export type TaskStageTasksSumAggregate = {
  id?: Maybe<Scalars["Float"]["output"]>;
  stageId?: Maybe<Scalars["Float"]["output"]>;
};
// ...
export type TaskUsersAggregateResponse = {
  avg?: Maybe<TaskUsersAvgAggregate>;
  count?: Maybe<TaskUsersCountAggregate>;
  groupBy?: Maybe<TaskUsersAggregateGroupBy>;
  max?: Maybe<TaskUsersMaxAggregate>;
  min?: Maybe<TaskUsersMinAggregate>;
  sum?: Maybe<TaskUsersSumAggregate>;
};
// ...
export type User = {
  avatarUrl?: Maybe<Scalars["String"]["output"]>;
  companies: UserCompaniesConnection;
  contacts: UserContactsConnection;
  createdAt: Scalars["DateTime"]["output"];
  createdBy?: Maybe<User>;
  deals: UserDealsConnection;
  email: Scalars["String"]["output"];
  events: UserEventsConnection;
  id: Scalars["ID"]["output"];
  jobTitle?: Maybe<Scalars["String"]["output"]>;
  name: Scalars["String"]["output"];
  phone?: Maybe<Scalars["String"]["output"]>;
  role: Role;
  tasks: UserTasksConnection;
  timezone?: Maybe<Scalars["String"]["output"]>;
  updatedAt: Scalars["DateTime"]["output"];
  updatedBy?: Maybe<User>;
};
// ...
export type UserCompaniesConnection = {
  /** Array of nodes. */
  nodes: Array<Company>;
  /** Paging information */
  pageInfo: OffsetPageInfo;
  /** Fetch total count of records */
  totalCount: Scalars["Int"]["output"];
};

export type UserConnection = {
  /** Array of nodes. */
  nodes: Array<User>;
  /** Paging information */
  pageInfo: OffsetPageInfo;
  /** Fetch total count of records */
  totalCount: Scalars["Int"]["output"];
};

export type UserContactsConnection = {
  /** Array of nodes. */
  nodes: Array<Contact>;
  /** Paging information */
  pageInfo: OffsetPageInfo;
  /** Fetch total count of records */
  totalCount: Scalars["Int"]["output"];
};

export type UserCreateInput = {
  email: Scalars["String"]["input"];
  jobTitle: Scalars["String"]["input"];
  name: Scalars["String"]["input"];
  phone: Scalars["String"]["input"];
  role: Role;
  timezone: Scalars["String"]["input"];
};

export type UserDealsConnection = {
  /** Array of nodes. */
  nodes: Array<Deal>;
  /** Paging information */
  pageInfo: OffsetPageInfo;
  /** Fetch total count of records */
  totalCount: Scalars["Int"]["output"];
};
// ...
export type UserEventsConnection = {
  /** Array of nodes. */
  nodes: Array<Event>;
  /** Paging information */
  pageInfo: OffsetPageInfo;
  /** Fetch total count of records */
  totalCount: Scalars["Int"]["output"];
};
// ...
export type UserTasksConnection = {
  /** Array of nodes. */
  nodes: Array<Task>;
  /** Paging information */
  pageInfo: OffsetPageInfo;
  /** Fetch total count of records */
  totalCount: Scalars["Int"]["output"];
};